Start redis server on boot
Unless I missed it, I did not see any documentation on telling the user to enable this server. Typically if someone doesn't know they should do this, then they most likely won't. Failing to run this command has caused a lot of unneeded issues due to the panel not being able to fetch the eggs since the redis-server isn't running.