Clarify something for Apache
It seems to be a weekly issue that for people with Apache, they have issues installing the panel because they follow the tutorial and then when they view it, it's displaying the actual PHP code. The notice I've inserted will help make users aware that when using Apache, they need to specific package for Apache that will make PHP run properly.