Mention additional NGINX reverse proxy config
For many web services (especially less complex ones), an NGINX reverse proxy does not need the config shown in this commit, so it is reasonable to assume that many users will be unaware of this config even if they already use NGINX as a reverse proxy. Since reverse proxy configuration is discussed, it only makes sense to mention everything that Pterodactyl requires. This config might be able to be trimmed down a little bit, but I personally do not have the time to test the importance of each line individually.