Added a note to help users with permissions for entrypoint.sh
I have added this note to the bottom of page to help user set the right permissions for their entrypoint.sh I had this problem, so I would just like to let other user know before they have to take to the Pterodactyl Discord Server. I just thought it would be a nice note. I don't mind if you change the way it is said or formatted, as long as it has the same meaning as what I wrote.