Change Depends
Changed PHP to 7.4 (`8` recommended) from 7.3+ as 7.3 is now on security updates until Dec 2021 -> EOL
Removed MariaDB from example to install depends, Added MySQL 8. To reduce the number of users getting errors when creating MySQL Users in MariaDB due to the `WITH mysql_native_password` part of the code block 
Removed apt update as this is automatically done when you add a repository. 


Note: This PR shouldn't be accepted until @DaneEveritt releases the next version of 1,X that will support PHP 8