PHP Repo Debian 10 Change (#372)
* PHP Repo

Personally, this is the only way I can get PHP installed nowadays. Used to work fine the old way before PHP 8 came out but now it just gives me the error php modules are not found and this fixes it for me.

* docs: add basic packages and change to curl

Includes basic packages to proceed in case they are not already in the distro source. Change wget to curl for consistency

Co-authored-by: Softwarenoob <admin@softwarenoob.com>