Add sudo where necessary
Added `sudo` to the beginning of any command that requires it to function when used with a user account with `sudo` privileges.

Without including `sudo` new users may be tempted to, or believe it's necessary, to switch to the root user. This definitely isn't something that should be done so best that the commands that need privilege escalation say so.