Update creating_ssl_certificates.md
Using the instructions above this code block, you create a Cloudflare API Token and not an API Key. The API token is the only variable you need to export before running the acme.sh command. You don't need to include an API Key or email and will work without it.